// be-cloud-crm-service-release/app/src/constants/enums/ticket.tags.enum.ts
export enum TicketTag {
  GENERAL_INFORMATION = 'GENERAL_INFORMATION',
  PAST_PAPERS = 'PAST_PAPERS',
  CAREERS = 'CAREERS',
  DOUBTASKEDDIRECTLY = 'DOUBTASKEDDIRECTLY',
  JUNIORS = 'JUNIORS',
  TECHNICAL_ISSUE = 'TECHNICAL_ISSUE',
  INCOMPLETE_CHAPTERS = 'INCOMPLETE_CHAPTERS',
  WRONG_ANSWER = 'WRONG_ANSWER',
  USER_FEEDBACK = 'USER_FEEDBACK',
  OTHER_BOARD = 'OTHER_BOARD',
  DS_MONETIZATION = 'DS_MONETIZATION',
  DS_SALE = 'DS_SALE',
  SIKKAY_INFO = 'SIKKAY_INFO',
  DS_SIKKAY_SALE = 'DS_SIKKAY_SALE',
  FB_DIRECT = 'FB_DIRECT',
  SOLUTION_DELAYED = 'SOLUTION_DELAYED',
  HOW_TO_USE = 'HOW_TO_USE',
  ONLINE_CLASSES_TEACHER = 'ONLINE_CLASSES_TEACHER',
  MDCAT = 'MDCAT',
  ECAT = 'ECAT',
  MDCAT_PURCHASE = 'MDCAT_PURCHASE',
  ECAT_PURCHASE = 'ECAT_PURCHASE',
  APPRECIATION = 'APPRECIATION',
  DS_INFO = 'DS_INFO',
  URDU_MEDIUM = 'URDU_MEDIUM',
  UNIVERSITY = 'UNIVERSITY',
  PASTPAPER_SALE = 'PASTPAPER_SALE',
  VIDEO_LECTURE_SALE = 'VIDEO_LECTURE_SALE',
  TESTS_SALE = 'TESTS_SALE',
  NEGATIVE = 'NEGATIVE',
  IRRELEVANT = 'IRRELEVANT',
  NO_CONVERSATION = 'NO_CONVERSATION',
  TEST_MONETIZATION = 'TEST_MONETIZATION',
  VIDEO_LECTURE = 'VIDEO_LECTURE',
  TESTS_INFO = 'TESTS_INFO',
  SUBJECT_REQUESTS = 'SUBJECT_REQUESTS',
  PAYMENT_ISSUE = 'PAYMENT_ISSUE',
  PAYMENT_INFO = 'PAYMENT_INFO',
  AFFORDABILITY = 'AFFORDABILITY',
  REQUESTED_FOR_NOTES = 'REQUESTED_FOR_NOTES',
  WHATSAPP_GROUP = 'WHATSAPP_GROUP',
  GUESS_PAPERS = 'GUESS_PAPERS',
  MARVEL_SUBMISSION = 'MARVEL_SUBMISSION',
  PROMO_CODE = 'PROMO_CODE',
  BUSINESS_QUERY = 'BUSINESS_QUERY',
  MATHS = 'MATHS',
  PHYSICS = 'PHYSICS',
  CHEMISTRY = 'CHEMISTRY',
  BIOLOGY = 'BIOLOGY',
  ENGLISH = 'ENGLISH',
  COMPUTER = 'COMPUTER',
  STATISTICS = 'STATISTICS',
  ACCOUNTING = 'ACCOUNTING',
  ECONOMICS = 'ECONOMICS',
  SENIOR = 'SENIOR',
  ARTS = 'ARTS',
  CLASS_1 = 'CLASS_1',
  CLASS_2 = 'CLASS_2',
  CLASS_3 = 'CLASS_3',
  CLASS_4 = 'CLASS_4',
  CLASS_5 = 'CLASS_5',
  CLASS_6 = 'CLASS_6',
  CLASS_7 = 'CLASS_7',
  CLASS_8 = 'CLASS_8',
  CLASS_9 = 'CLASS_9',
  CLASS_10 = 'CLASS_10',
  CLASS_11 = 'CLASS_11',
  CLASS_12 = 'CLASS_12',
  EXAM_BUNDLE = 'EXAM_BUNDLE',
  OTP_ISSUE = 'OTP_ISSUE',
  OTP_SENT = 'OTP_SENT',
  REVISION_BUNDLE = 'REVISION_BUNDLE',
  CJ_INFO_9_10 = 'CJ_INFO_9_10',
  CJ_INFO_11_12 = 'CJ_INFO_11_12',
  DHAMAKAYDAR_OFFER = 'DHAMAKAYDAR_OFFER',
  MEC_INFO = 'MEC_INFO',
}